Add snapshot download support

+++ scripts/start_mainnet_fork.sh
@@ -3,7 +3,7 @@
 neutrond tendermint unsafe-reset-all --home /opt/neutron/data
 
 CUSTOM_SCRIPT_PATH=/opt/neutron/custom/config.sh
-SNAPSHOT_DOWNLOAD_URL="http://172.17.0.1:8080"
+SNAPSHOT_DOWNLOAD_URL="https://raw-snapshots.neutron.org"
 
 if [ ! -d "/opt/neutron/data_backup" ]; then    
     echo "Previous state backup not found, starting from genesis..."
@@ -12,8 +12,25 @@ if [ ! -d "/opt/neutron/data_backup" ]; then
         echo "Snapshot not found, downloading it from snapshot service..."
 
         METADATA=$(curl -s $SNAPSHOT_DOWNLOAD_URL/.metadata.json)
+        if [ -z "$METADATA" ]; then
+            echo "Snapshot metadata not found, aborting..."
+            exit 1
+        fi
+
+        if ! echo $METADATA | jq empty 2>/dev/null; then
+            echo "Metadata file is not valid json."
+            exit 1
+        fi
+
+        if ! echo $METADATA | jq -e 'has("snapshot_path")' > /dev/null; then
+            echo "Wrong metadata file type, aborting..."
+            exit 1
+        fi        
+
         SNAPSHOT_ARCHIVE=$(echo "$METADATA" | jq -r .snapshot_path)
         SNAPSHOT_NAME=$(echo "$METADATA" | jq -r .snapshot_name)
+        echo "Downloading $SNAPSHOT_ARCHIVE..."
+        echo "Snapshot name: $SNAPSHOT_NAME"
         wget ${SNAPSHOT_DOWNLOAD_URL}/$SNAPSHOT_ARCHIVE -O /opt/neutron/snapshot/$SNAPSHOT_ARCHIVE
         gunzip -f /opt/neutron/snapshot/$SNAPSHOT_ARCHIVE 
         mv -f /opt/neutron/snapshot/$SNAPSHOT_NAME /opt/neutron/snapshot/snapshot.json
